pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
authors = [
    {name = "SK-415", email = "2967923486@qq.com"},
]
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.8.1,<4.0"
dependencies = [
    "click>=8.1.3",
    "httpx>=0.24.0",
    "nonebot-adapter-onebot>=2.2.3",
    "nonebot-plugin-apscheduler>=0.2.0",
    "uvicorn[standard]>=0.23.2",
    "fastapi>=0.104.0",
    "nonebot2>=2.1.1",
    "playwright>=1.33.0",
    "pydantic>=1.10.7",
    "python-dotenv>=1.0.0",
    "tortoise-orm[asyncpg]>=0.19.3",
    "bilireq>=0.2.11",
    "packaging>=23.1",
    "nonebot-plugin-guild-patch>=0.2.3",
    "msvc-runtime>=14.34.31931; sys_platform == \"win32\"",
    "qrcode>=7.4.2",
    "pillow>=9.2.0",
    "nonebot-plugin-htmlrender>=0.2.2",
    "pytest-playwright>=0.4.3",
    "PyExecJS>=1.5.1",
    "nb-cli>=1.2.6",
    "skia-python>=87.5",
    "emoji>=2.6.0",
    "dynamicadaptor>=0.4.5",
    "bilirpc>=0.1.3",
    "nonebot-plugin-sentry>=0.5.0",
    "nonebot-plugin-picstatus>=0.5.3",
]
name = "haruka-bot"
version = "1.5.3.post1"
description = "Push dynamics and live informations from bilibili to QQ. Based on nonebot2."
readme = "README.md"
keywords = ["nonebot", "nonebot2", "qqbot", "bilibili", "bot"]

[project.urls]
homepage = "https://github.com/SK-415/HarukaBot"
repository = "https://github.com/SK-415/HarukaBot/tree/master/src/plugins/haruka_bot"
documentation = "https://github.com/SK-415/HarukaBot#readme"

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.7.0",
    "pre-commit>=3.3.3",
    "flake8>=6.1.0",
    "nonebot-plugin-sentry>=0.4.1",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_sentry", "nonebot_plugin_htmlrender", "src", "nonebot_plugin_picstatus"]
plugin_dirs = []
builtin_plugins = []

[project.scripts]
hb = "haruka_bot.__main__:main"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

line-length = 88
target-version = "py38"

[tool.ruff.lint]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["nonebot", "tests"]
extra_standard_library = ["typing_extensions"]